requirements.txt
streamlit
opencv-python-headless
tensorflow
sympy
numpy
Pillow
